fix: close socket if player exits before opponent joins
Also changed return type of wait_for_game_start() from () to Result, so that it won't panic if the connection couldn't be made (a cause can be the player going back to Main Menu).
(WIP) Resolve bug: After the player exits before an opponent joins, if he selects Normal game or Multiplayer in the Main Menu, it won't work anymore.